FLPATH-905: make the notifications icon active in dynamic plugin
