+++ plugins/notifications/src/components/NotificationsActiveIcon/index.ts
@@ -1,0 +1,1 @@
+export * from './NotificationsActiveIcon';
