+++ plugins/notifications/src/constants.ts
@@ -1,3 +1,2 @@
-export const NOTIFICATIONS_ROUTE = 'notifications';
-
 export const DebounceDelayMs = 1000;
+export const DefaultPollingIntervalMs = 5 * 1000;
