+++ plugins/notifications/src/routes.ts
@@ -1,7 +1,5 @@
 import { createRouteRef } from '@backstage/core-plugin-api';
 
-import { NOTIFICATIONS_ROUTE } from './constants';
-
-export const rootRouteRef = createRouteRef({
-  id: NOTIFICATIONS_ROUTE,
+export const notificationsRootRouteRef = createRouteRef({
+  id: 'notifications',
 });
